+++ src/test/testData/project/magento2/app/code/Foo/Bar/view/frontend/requirejs-config.js
@@ -6,5 +6,17 @@ var config = {
     },
     paths: {
         'testFile2': 'Foo_Bar/js/file2'
+    },
+    config: {
+        mixins: {
+            'Foo_Bar/js/file': {
+                'Foo_Bar/js/file-mixin': true
+            },
+            'Foo_Bar/js/file2': {
+                'Foo_Bar/js/object-mixin': true,
+                'Foo_Bar/js/widget-mixin': true,
+                'Foo_Bar/js/function-mixin': true
+            }
+        }
     }
 }