+++ src/test/testData/project/magento2/app/code/Foo/Bar/view/frontend/web/js/object-mixin.js
@@ -1,0 +1,8 @@
+define(function () {
+    'use strict';
+
+    return function (target) {
+        target.mixed = true;
+        return target;
+    };
+});